+++ src/lightning_app/CHANGELOG.md
@@ -47,6 +47,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Fixed catimage import ([#15712](https://github.com/Lightning-AI/lightning/pull/15712))
 - Parse all lines in app file looking for shebangs to run commands ([#15714](https://github.com/Lightning-AI/lightning/pull/15714))
 
+- Fixed setting property to the LightningFlow ([#15750](https://github.com/Lightning-AI/lightning/pull/15750))
+
+
 
 ## [1.8.1] - 2022-11-10